Rename package to sink
