+++ server/collector/src/main/java/org/bithon/server/collector/sink/kafka/KafkaMetricSink.java
@@ -19,9 +19,9 @@ package org.bithon.server.collector.sink.kafka;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.kafka.clients.producer.ProducerRecord;
-import org.bithon.server.metric.handler.IMetricMessageSink;
+import org.bithon.server.metric.sink.IMetricMessageSink;
 import org.bithon.server.common.utils.collection.CloseableIterator;
-import org.bithon.server.metric.handler.MetricMessage;
+import org.bithon.server.metric.sink.MetricMessage;
 import org.springframework.kafka.core.KafkaTemplate;
 
 import java.nio.charset.StandardCharsets;
