+++ server/collector/src/main/java/org/bithon/server/collector/source/brpc/BrpcCollectorConfig.java
@@ -20,18 +20,18 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableMap;
 import lombok.Data;
 import org.apache.kafka.common.serialization.StringSerializer;
-import org.bithon.server.event.handler.IEventMessageSink;
-import org.bithon.server.metric.handler.IMessageSink;
-import org.bithon.server.metric.handler.IMetricMessageSink;
-import org.bithon.server.tracing.handler.ITraceMessageSink;
+import org.bithon.server.event.sink.IEventMessageSink;
+import org.bithon.server.metric.sink.IMessageSink;
+import org.bithon.server.metric.sink.IMetricMessageSink;
+import org.bithon.server.tracing.sink.ITraceMessageSink;
 import org.bithon.server.collector.sink.kafka.KafkaEventSink;
 import org.bithon.server.collector.sink.kafka.KafkaMetricSink;
 import org.bithon.server.collector.sink.kafka.KafkaTraceSink;
-import org.bithon.server.event.handler.LocalEventSink;
-import org.bithon.server.metric.handler.LocalMetricSink;
-import org.bithon.server.metric.handler.LocalSchemaMetricSink;
-import org.bithon.server.tracing.handler.LocalTraceSink;
-import org.bithon.server.metric.handler.SchemaMetricMessage;
+import org.bithon.server.event.sink.LocalEventSink;
+import org.bithon.server.metric.sink.LocalMetricSink;
+import org.bithon.server.metric.sink.LocalSchemaMetricSink;
+import org.bithon.server.tracing.sink.LocalTraceSink;
+import org.bithon.server.metric.sink.SchemaMetricMessage;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.ApplicationContext;
