+++ server/collector/src/main/java/org/bithon/server/collector/source/brpc/BrpcCollectorStarter.java
@@ -27,10 +27,10 @@ import org.bithon.component.brpc.channel.ServerChannel;
 import org.bithon.server.collector.cmd.api.CommandService;
 import org.bithon.server.collector.setting.AgentSettingService;
 import org.bithon.server.collector.setting.BrpcSettingFetcher;
-import org.bithon.server.event.handler.IEventMessageSink;
-import org.bithon.server.metric.handler.IMessageSink;
-import org.bithon.server.metric.handler.IMetricMessageSink;
-import org.bithon.server.tracing.handler.ITraceMessageSink;
+import org.bithon.server.event.sink.IEventMessageSink;
+import org.bithon.server.metric.sink.IMessageSink;
+import org.bithon.server.metric.sink.IMetricMessageSink;
+import org.bithon.server.tracing.sink.ITraceMessageSink;
 import org.springframework.beans.BeansException;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.ApplicationContext;
