+++ server/collector/src/main/java/org/bithon/server/collector/source/brpc/BrpcEventCollector.java
@@ -20,9 +20,9 @@ import lombok.extern.slf4j.Slf4j;
 import org.bithon.agent.rpc.brpc.BrpcMessageHeader;
 import org.bithon.agent.rpc.brpc.event.BrpcEventMessage;
 import org.bithon.agent.rpc.brpc.event.IEventCollector;
-import org.bithon.server.event.handler.IEventMessageSink;
+import org.bithon.server.event.sink.IEventMessageSink;
 import org.bithon.server.common.utils.collection.CloseableIterator;
-import org.bithon.server.event.handler.EventMessage;
+import org.bithon.server.event.sink.EventMessage;
 
 import java.util.Collections;
 import java.util.Iterator;
