+++ server/collector/src/main/java/org/bithon/server/collector/source/brpc/BrpcMetricCollector.java
@@ -34,8 +34,8 @@ import org.bithon.agent.rpc.brpc.metrics.BrpcSqlMetricMessage;
 import org.bithon.agent.rpc.brpc.metrics.BrpcThreadPoolMetricMessage;
 import org.bithon.agent.rpc.brpc.metrics.BrpcWebServerMetricMessage;
 import org.bithon.agent.rpc.brpc.metrics.IMetricCollector;
-import org.bithon.server.metric.handler.IMessageSink;
-import org.bithon.server.metric.handler.IMetricMessageSink;
+import org.bithon.server.metric.sink.IMessageSink;
+import org.bithon.server.metric.sink.IMetricMessageSink;
 import org.bithon.server.common.utils.ReflectionUtils;
 import org.bithon.server.common.utils.collection.CloseableIterator;
 import org.bithon.server.metric.DataSourceSchema;
@@ -47,8 +47,8 @@ import org.bithon.server.metric.aggregator.spec.LongMinMetricSpec;
 import org.bithon.server.metric.aggregator.spec.LongSumMetricSpec;
 import org.bithon.server.metric.dimension.IDimensionSpec;
 import org.bithon.server.metric.dimension.StringDimensionSpec;
-import org.bithon.server.metric.handler.MetricMessage;
-import org.bithon.server.metric.handler.SchemaMetricMessage;
+import org.bithon.server.metric.sink.MetricMessage;
+import org.bithon.server.metric.sink.SchemaMetricMessage;
 import org.springframework.util.CollectionUtils;
 
 import java.util.Iterator;
