+++ server/collector/src/main/java/org/bithon/server/collector/source/thrift/ThriftCollectorStarter.java
@@ -33,9 +33,9 @@ import org.bithon.agent.rpc.thrift.service.setting.SettingService;
 import org.bithon.agent.rpc.thrift.service.trace.ITraceCollector;
 import org.bithon.server.collector.setting.AgentSettingService;
 import org.bithon.server.collector.setting.SettingServiceThriftImpl;
-import org.bithon.server.event.handler.IEventMessageSink;
-import org.bithon.server.metric.handler.IMetricMessageSink;
-import org.bithon.server.tracing.handler.ITraceMessageSink;
+import org.bithon.server.event.sink.IEventMessageSink;
+import org.bithon.server.metric.sink.IMetricMessageSink;
+import org.bithon.server.tracing.sink.ITraceMessageSink;
 import org.springframework.beans.BeansException;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.ApplicationContext;
