+++ server/collector/src/main/java/org/bithon/server/collector/source/thrift/ThriftMetricCollector.java
@@ -32,8 +32,8 @@ import org.bithon.agent.rpc.thrift.service.metric.message.ThreadPoolMetricMessag
 import org.bithon.agent.rpc.thrift.service.metric.message.WebServerMetricMessage;
 import org.bithon.server.common.utils.ReflectionUtils;
 import org.bithon.server.common.utils.collection.CloseableIterator;
-import org.bithon.server.metric.handler.IMetricMessageSink;
-import org.bithon.server.metric.handler.MetricMessage;
+import org.bithon.server.metric.sink.IMetricMessageSink;
+import org.bithon.server.metric.sink.MetricMessage;
 import org.springframework.util.CollectionUtils;
 
 import java.util.Iterator;
