+++ server/sink/src/main/java/org/bithon/server/collector/source/kafka/KafkaMetricCollector.java
@@ -16,9 +16,9 @@
 
 package org.bithon.server.collector.source.kafka;
 
-import org.bithon.server.metric.handler.LocalMetricSink;
+import org.bithon.server.metric.sink.LocalMetricSink;
 import org.bithon.server.common.utils.collection.CloseableIterator;
-import org.bithon.server.metric.handler.MetricMessage;
+import org.bithon.server.metric.sink.MetricMessage;
 
 /**
  * Kafka collector that is connecting to KafkaMetricSink
