+++ server/sink/src/main/java/org/bithon/server/collector/source/kafka/KafkaTraceCollector.java
@@ -16,9 +16,9 @@
 
 package org.bithon.server.collector.source.kafka;
 
-import org.bithon.server.tracing.handler.LocalTraceSink;
+import org.bithon.server.tracing.sink.LocalTraceSink;
 import org.bithon.server.common.utils.collection.CloseableIterator;
-import org.bithon.server.tracing.handler.TraceSpan;
+import org.bithon.server.tracing.sink.TraceSpan;
 
 /**
  * @author frank.chen021@outlook.com
