+++ server/sink/src/main/java/org/bithon/server/event/sink/IEventMessageSink.java
@@ -1,8 +1,7 @@
-package org.bithon.server.event.handler;
+package org.bithon.server.event.sink;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import org.bithon.server.common.utils.collection.CloseableIterator;
-import org.bithon.server.event.handler.EventMessage;
 
 /**
  * @author Frank Chen
