+++ server/sink/src/main/java/org/bithon/server/metric/sink/HttpOutgoingMetricMessageHandler.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.metric.handler;
+package org.bithon.server.metric.sink;
 
 import lombok.extern.slf4j.Slf4j;
 import org.bithon.server.common.service.UriNormalizer;
