+++ server/sink/src/main/java/org/bithon/server/metric/sink/IMessageSink.java
@@ -14,27 +14,11 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.event.handler;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.Map;
+package org.bithon.server.metric.sink;
 
 /**
- * @author frank.chen021@outlook.com
- * @date 2021/2/14 4:11 下午
+ * @author frankchen
  */
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class EventMessage {
-    private String appName;
-    private String instanceName;
-    private Long timestamp;
-    private String type;
-    private Map<String, String> args;
+public interface IMessageSink<MSG> {
+    void process(String messageType, MSG message);
 }
