+++ server/sink/src/main/java/org/bithon/server/tracing/storage/ITraceReader.java
@@ -16,7 +16,7 @@
 
 package org.bithon.server.tracing.storage;
 
-import org.bithon.server.tracing.handler.TraceSpan;
+import org.bithon.server.tracing.sink.TraceSpan;
 
 import java.util.List;
