+++ server/web-service/src/main/java/org/bithon/server/web/service/tracing/api/TraceSpanBo.java
@@ -17,7 +17,7 @@
 package org.bithon.server.web.service.tracing.api;
 
 import lombok.Data;
-import org.bithon.server.tracing.handler.TraceSpan;
+import org.bithon.server.tracing.sink.TraceSpan;
 
 import java.util.ArrayList;
 import java.util.List;
