+++ server/web-service/src/main/java/org/bithon/server/web/service/tracing/service/TraceService.java
@@ -18,7 +18,7 @@ package org.bithon.server.web.service.tracing.service;
 
 import org.bithon.server.web.service.tracing.api.TraceMap;
 import org.bithon.server.web.service.tracing.api.TraceSpanBo;
-import org.bithon.server.tracing.handler.TraceSpan;
+import org.bithon.server.tracing.sink.TraceSpan;
 import org.bithon.server.tracing.storage.ITraceReader;
 import org.bithon.server.tracing.storage.ITraceStorage;
 import org.springframework.beans.BeanUtils;
